Added AdaptiveLayout macro
Added the declaration and implementation of the macro.

Spent a day or two realizing we can in fact just add code, and the macro packages will parse that into the syntax tree it needs. No need to dig around in the syntax or debug to find the right syntax types, and the source is more readable. Then another day swearing about the macro not working until I noticed the stringify rename missed the list of macros provided to the unit test.

Expansion works locally in Bookmind, unit test passes, docs look great. We'll need a README for the public repo, maybe some changes to get that working remotely for Bookmind, then I think we have a v1. If I find any other layout code suitable for macros we'll see about a v2.